// frontOffice/src/app/models/user.model.ts
export interface User {
    _id: String|undefined;
    name: String | undefined;
    email: String | undefined;
    password: String | undefined;
    role: String | undefined;
    phoneNumber: String | undefined;
    address: String | undefined;
    country: String | undefined;
    points: Number | undefined;
    dateOfBirthday: Date | undefined;
    profileImage: any;
}